requirements.txt
asgiref==3.8.1
Django==5.1.8
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
Faker==37.1.0
inflection==0.5.1
Markdown==3.7
packaging==24.2
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
validate-docbr==1.10.0
